UE-160860 - Fix UnrealEd Module truncation warnings Part 3 of N
#jira UE-160860
#preflight 6348a349204d7f312b70632c
#rb andrew.davidson

[CL 22602879 by dave jones2 in ue5-main branch]

+++ Engine/Source/Editor/UnrealEd/Private/AutoReimport/AutoReimportManager.cpp
@@ -758,7 +758,7 @@ TOptional<ECurrentState> FAutoReimportManager::ProcessDeletions()
 		Monitor->ExtractAssetsToDelete(AssetsToDelete);
 	}
 
-	FeedbackContextOverride->MainTask->EnterProgressFrame(TotalWork);
+	FeedbackContextOverride->MainTask->EnterProgressFrame(static_cast<float>(TotalWork));
 
 	if (AssetsToDelete.Num() > 0)
 	{
